+++ hutool-core/src/main/java/cn/hutool/core/bean/NullWrapperBean.java
@@ -2,19 +2,28 @@ package cn.hutool.core.bean;
 
 /**
  * 为了解决反射过程中,需要传递null参数,但是会丢失参数类型而设立的包装类
+ *
+ * @param <T> Null值对应的类型
+ * @author Lillls
+ * @since 5.5.0
  */
-public class NullWrapperBean {
+public class NullWrapperBean<T> {
 
-    private final Class<?> mClasses;
+	private final Class<T> clazz;
 
-    /**
-     * @param classes null的类型
-     */
-    public NullWrapperBean(Class<?> classes) {
-        this.mClasses = classes;
-    }
+	/**
+	 * @param clazz null的类型
+	 */
+	public NullWrapperBean(Class<T> clazz) {
+		this.clazz = clazz;
+	}
 
-    public Class<?> getClasses() {
-        return mClasses;
-    }
+	/**
+	 * 获取null值对应的类型
+	 *
+	 * @return 类型
+	 */
+	public Class<T> getWrappedClass() {
+		return clazz;
+	}
 }