+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * 【core   】     修复Ipv4Util.longToIp的问题
 * 【poi    】     修复Excel07SaxReader读取公式的错误的问题（issue#I23VFL@Gitee）
 * 【http   】     修复HttpUtil.isHttp判断问题（pr#1208@Github）
+* 【http   】     修复Snowflake时间回拨导致ID重复的bug（issue#1206@Github）
 
 -------------------------------------------------------------------------------------------------------------